Add control of window transformation and selecting of polygon

+++ bones/include/Point.h
@@ -28,11 +28,11 @@ class Point {
 	}
 	
 
-	int getX() const {
+	double getX() const {
 		return x;
 	}
 
-	int getY() const {
+	double getY() const {
 		return y;
 	}
